Airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
minio
langchain
langchain-core
langchain_community
langchain-text-splitters
chromadb
tiktoken
pypdf
openai
pdfplumber
python-docx